fix(view): removed Card with force directed menu from MainWorkArea.kt to ForceDirectedMenu.kt
